requirements.txt
python-telegram-bot
pymongo
python-dotenv
Pillow
pytesseract
opencv-python
cryptography
